Only move paritaly update to AnomalyFunctionResource at this time
