+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/dashboard/resources/AnomalyResource.java
@@ -245,7 +245,6 @@ public class AnomalyResource {
    return response;
  }
 
-
   /************* CRUD for anomaly functions of collection **********************************************/
   // View all anomaly functions
   @GET
@@ -442,36 +441,6 @@ public class AnomalyResource {
     anomalyFunctionDAO.update(anomalyFunctionSpec);
     return Response.ok(id).build();
   }
-
-  // Partially update anomaly function
-  @POST
-  @Path("/anomaly-function/update/properties")
-  public Response updateAnomalyFunctionProperties (@QueryParam("id") @NotNull Long id,
-      @QueryParam("config") @NotNull String propertiesJson) {
-    if(id == null || anomalyFunctionDAO.findById(id) == null) {
-      String msg = "Unable to update function properties. " + id + " doesn't exist";
-      LOG.warn(msg);
-      return Response.status(Response.Status.BAD_REQUEST).entity(msg).build();
-    }
-
-    if(StringUtils.isNotBlank(propertiesJson)) {
-      Map<String, String> configs = Collections.emptyMap();
-      try {
-        configs = OBJECT_MAPPER.readValue(propertiesJson, Map.class);
-      } catch (IOException e) {
-        String msg = "Unable to parse json string " + propertiesJson + " for function " + id;
-        LOG.error(msg);
-        return Response.status(Response.Status.BAD_REQUEST).entity(msg).build();
-      }
-      AnomalyFunctionDTO anomalyFunction = anomalyFunctionDAO.findById(id);
-      anomalyFunction.updateProperties(configs);
-      anomalyFunctionDAO.update(anomalyFunction);
-    }
-    String msg = "Successfully update properties for function " + id + " with " + propertiesJson;
-    LOG.info(msg);
-    return Response.ok(id).build();
-  }
-
 
   /**
    * Apply an autotune configuration to an existing function
@@ -704,105 +673,6 @@ public class AnomalyResource {
     }
 
       return idList;
-  }
-
-  // Activate anomaly function
-  @POST
-  @Path("/anomaly-function/activate")
-  public Response activateAnomalyFunction(@NotNull @QueryParam("functionId") Long id) {
-    if (id == null) {
-      return Response.status(Response.Status.BAD_REQUEST).build();
-    }
-    toggleFunctionById(id, true);
-    return Response.ok(id).build();
-  }
-
-  // batch activate and deactivate anomaly functions
-  @POST
-  @Path("/anomaly-functions/activate")
-  public String activateFunction(@QueryParam("functionIds") String functionIds) {
-    toggleFunctions(functionIds, true);
-    return functionIds;
-  }
-
-  @POST
-  @Path("/anomaly-functions/deactivate")
-  public String deactivateFunction(@QueryParam("functionIds") String functionIds) {
-    toggleFunctions(functionIds, false);
-    return functionIds;
-  }
-
-  /**
-   * toggle anomaly functions to active and inactive
-   *
-   * @param functionIds string comma separated function ids, ALL meaning all functions
-   * @param isActive boolean true or false, set function as true or false
-   */
-  private void toggleFunctions(String functionIds, boolean isActive) {
-    List<Long> functionIdsList = new ArrayList<>();
-
-    // can add tokens here to activate and deactivate all functions for example
-    // functionIds == {SPECIAL TOKENS} --> functionIdsList = anomalyFunctionDAO.findAll()
-
-    if (StringUtils.isNotBlank(functionIds)) {
-      String[] tokens = functionIds.split(",");
-      for (String token : tokens) {
-        functionIdsList.add(Long.valueOf(token));  // unhandled exception is expected
-      }
-    }
-
-    for (long id : functionIdsList) {
-      toggleFunctionById(id, isActive);
-    }
-  }
-
-  private void toggleFunctionById(long id, boolean isActive) {
-    AnomalyFunctionDTO anomalyFunctionSpec = anomalyFunctionDAO.findById(id);
-    anomalyFunctionSpec.setActive(isActive);
-    anomalyFunctionDAO.update(anomalyFunctionSpec);
-  }
-
-  // Delete anomaly function
-  @DELETE
-  @Path("/anomaly-function/delete")
-  public Response deleteAnomalyFunctions(@NotNull @QueryParam("id") Long id,
-      @QueryParam("functionName") String functionName)
-      throws IOException {
-
-    if (id == null) {
-      throw new IllegalArgumentException("id is a required query param");
-    }
-
-    // call endpoint to shutdown if active
-    AnomalyFunctionDTO anomalyFunctionSpec = anomalyFunctionDAO.findById(id);
-    if (anomalyFunctionSpec == null) {
-      throw new IllegalStateException("No anomalyFunctionSpec with id " + id);
-    }
-
-    // delete dependent entities
-    // email config mapping
-    List<EmailConfigurationDTO> emailConfigurations = emailConfigurationDAO.findByFunctionId(id);
-    for (EmailConfigurationDTO emailConfiguration : emailConfigurations) {
-      emailConfiguration.getFunctions().remove(anomalyFunctionSpec);
-      emailConfigurationDAO.update(emailConfiguration);
-    }
-
-    // raw result mapping
-    List<RawAnomalyResultDTO> rawResults =
-        rawAnomalyResultDAO.findAllByTimeAndFunctionId(0, System.currentTimeMillis(), id);
-    for (RawAnomalyResultDTO result : rawResults) {
-      rawAnomalyResultDAO.delete(result);
-    }
-
-    // merged anomaly mapping
-    List<MergedAnomalyResultDTO> mergedResults = anomalyMergedResultDAO.findByFunctionId(id, true);
-    for (MergedAnomalyResultDTO result : mergedResults) {
-      anomalyMergedResultDAO.delete(result);
-    }
-
-    // delete from db
-    anomalyFunctionDAO.deleteById(id);
-    return Response.noContent().build();
   }
 
   /************ Anomaly Feedback **********/
